+++ src/fastapibootcamp/handlers/external.py
@@ -3,17 +3,41 @@
 from typing import Annotated
 
 from fastapi import APIRouter, Depends
+from pydantic import BaseModel, Field
 from safir.dependencies.logger import logger_dependency
+from safir.metadata import Metadata as SafirMetadata
 from safir.metadata import get_metadata
 from structlog.stdlib import BoundLogger
 
 from ..config import config
-from ..models import Index
 
 __all__ = ["get_index", "external_router"]
 
 external_router = APIRouter()
 """FastAPI router for all external handlers."""
+
+
+# In the default template there's a "models" module that defines the Pydantic
+# models. For this router, we're going to co-locate models and path operation
+# function in the same module to make the demo easier to follow. For a real
+# application, I recommend keeping models in their own module, but instead of
+# a single root-level "models" module, keep the API models next to the
+# handlers, and have internal models elsewhere in the "domain" and "storage"
+# interface subpackages.
+
+
+class Index(BaseModel):
+    """Metadata returned by the external root URL of the application.
+
+    Notes
+    -----
+    As written, this is not very useful. Add additional metadata that will be
+    helpful for a user exploring the application, or replace this model with
+    some other model that makes more sense to return from the application API
+    root.
+    """
+
+    metadata: SafirMetadata = Field(..., title="Package metadata")
 
 
 @external_router.get(
